pyproject.toml
[build-system]
requires = ["setuptools >= 75.0"]
build-backend = "setuptools.build_meta"

[project]
name = "todo_api"
version = "0.1.0"
dependencies = [
    "fastapi[standard] >= 0.114",
    "pyjwt >= 2.9",
    "passlib[bcrypt] >= 1.7",
    "psycopg[binary, pool] >= 3.2",
    "SQLAlchemy >= 2.0",
]
requires-python = ">=3.12"

[project.optional-dependencies]
testing = [
    "ruff >= 0.6",
    "pytest >= 8.3",
    "pytest-cov >= 5.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"

[tool.pytest.ini_options]
addopts = "--cov=todo_api"
pythonpath = [
    ".", "src",
]
testpaths = [
    "tests",
]
